+++ resources/views/toolbar/partials/dropdown-checkbox.blade.php
@@ -6,7 +6,7 @@
     /** @var \Entelechy\Architect\Toolbar\Items\Dropdown\DropdownCheckbox $dropdownItem */
     if (!$this->can($dropdownItem->getPermission())) return;
 
-    $compoundKey = $parentDropKey . '.' . $dropdownItem->key();
+    $compoundKey = $parentDropKey . '.' . $dropdownItem->getKey();
     $isChecked   = $checkboxValues[$compoundKey] ?? $dropdownItem->getDefault();
     $persist     = $dropdownItem->getPersist();
     $lsKey       = $persist === 'local'